Scale motion model by dT
