+++ src/localization/motion_model.py
@@ -29,7 +29,6 @@ class MotionModel:
         #X_k1: sin(Theta_(k-1)1)*dX1 + cos(Theta_(k-1)1) + Y_{k-1}1
         #Theta_k1: acos(cos(Theta_(k-1)1)*cos(dX1) - sin(Theta_(k-1)1)*sin(dX1)) = acos(cos(Theta_(k-1)1  + dX1)) = Theta_(k-1)1  + dX1
         """
-        
         # Evaluate above Derivation
         particles[:,0] = np.cos(particles[:,2])*odometry[0] - np.sin(particles[:,2])*odometry[1] + particles[:,0]
         particles[:,1] = np.sin(particles[:,2])*odometry[0] + np.cos(particles[:,2])*odometry[1] + particles[:,1]
